+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 #### 5.0.0-beta.14
-* Fix #340 — Pass value and formatted values 
+* Fix #340 — Pass value and formatted values
+* Fix #339 — Handle more common/known types in the Json serialiser
 
 #### 5.0.0-beta.13
 * Fix inverted if-statement in printDotOnOverflow